Module cribbage::domain::prelude

Re-exports

pub use crate::domain::*;

Structs

Card

A playing card.

Play

PlayState

Player

Rank

The rank of a Card. Ace(1) to King(13).

Score

Value

The value of a Card. 1 to 10.

Enums

Error

Game

The game state, waiting for opponent, discarding, playing, scoring, finished.

Constants

CARDS_DEALT_PER_HAND

Six [Card]s are dealt to each Player by the current dealer.

CARDS_DISCARDED_TO_CRIB

CARDS_KEPT_PER_HAND

[Player]s discard two cards each into the [Crib] leaving four for [Score]ing and [Plays].

CARDS_REQUIRED_IN_CRIB

Each [Player] discarding two [Card]s to the [Crib] will mean four [Card]s end up there.

NUMBER_OF_PLAYERS_IN_GAME

Cribbage can be a two, three or (at a push) four player game. This implementation is for two players.

PLAY_TARGET

Each [Plays.Play] cannot have a running total of more than 31.

WINNING_SCORE

Short games can play to 61, but it is normal to play to 121, as in this implementation.

Type Aliases

Crib

Cut

Deck

Hand